requirements.txt
google_api_python_client==2.139.0
hachoir==3.3.0
piexif==1.1.3
Pillow==10.4.0
pillow_heif==0.18.0
protobuf==3.20.3
selenium==4.23.1
